+++ android/app/src/main/java/com/didroom/wallet/OldAndroidVersionException.java
@@ -1,7 +1,8 @@
-package didroom.forkbomb.eu;
+package com.didroom.wallet;
 
 public class OldAndroidVersionException extends Exception {
     OldAndroidVersionException() {
         super("Android version less than minimum supported");
     }
 }
+